requirements.txt
pyperclip==1.8.2
PyQt5
google-cloud-translate
